+++ src/tripCard/index.tsx
@@ -7,16 +7,21 @@ import ComfortIcon from 'icon/comfortIcon'
 import LightningIcon from 'icon/lightningIcon'
 import LadyIcon from 'icon/ladyIcon'
 import Itinerary from 'itinerary'
+import Item from '_utils/item'
 import Text, { TextDisplayType } from 'text'
 import style from 'tripCard/style'
+import { color } from '_utils/branding'
+
+interface User {
+  avatarUrl: string
+  firstName: string
+}
 
 export interface TripCardProps {
   href: string | JSX.Element
   itinerary: Place[]
-  driver: {
-    avatarUrl: string
-    firstName: string
-  }
+  driver?: User
+  passengers?: User[]
   price: string
   flags?: {
     ladiesOnly?: boolean
@@ -31,6 +36,12 @@ export interface TripCardProps {
   metaUrl: string
   highlighted?: string
   className?: Classcat.Class
+  statusInformation?: {
+    icon: JSX.Element,
+    text: string,
+    highlighted?: boolean,
+  }
+  badge?: string
 }
 
 const TripCard = ({
@@ -38,11 +49,14 @@ const TripCard = ({
   href,
   itinerary,
   driver,
+  passengers = [],
   price,
   flags = {},
   titles = {},
   highlighted = '',
   metaUrl,
+  statusInformation = null,
+  badge = null,
 }: TripCardProps) => {
   const departure = itinerary[0]
   const arrival = itinerary[itinerary.length - 1]
@@ -69,7 +83,10 @@ const TripCard = ({
 
   return (
     <li
-      className={cc(['kirk-tripCard', { 'kirk-tripCard--highlighted': !!highlighted }, className])}
+      className={cc(['kirk-tripCard', {
+        'kirk-tripCard--highlighted': !!highlighted,
+        'kirk-tripCard--with-badge': badge && badge.length,
+      }, className])}
       itemScope
       itemType="http://schema.org/Event"
     >
@@ -82,6 +99,28 @@ const TripCard = ({
           <meta itemProp="startDate" content={departure.isoDate} />
           <meta itemProp="endDate" content={arrival.isoDate} />
 
+          {badge && (
+            <Text
+              className="kirk-tripCard-badge"
+              textColor={color.white}>
+              {badge}
+            </Text>
+          )}
+
+          {statusInformation && (
+            <div className="kirk-tripCard-top">
+              <Item
+                className="kirk-tripCard-top-item"
+                leftAddon={React.cloneElement(statusInformation.icon, {
+                  iconColor: statusInformation.highlighted ? color.primary : color.icon
+                })}
+                leftTitle={statusInformation.text}
+                leftTitleDisplay={TextDisplayType.BODY}
+                highlighted={statusInformation.highlighted}
+              />
+            </div>
+          )}
+
           <div className="kirk-tripCard-main">
             <Itinerary className="kirk-tripCard-itinerary" places={itinerary} />
             <Text className="kirk-tripCard-price" display={TextDisplayType.TITLESTRONG}>
@@ -89,12 +128,23 @@ const TripCard = ({
             </Text>
           </div>
           <div className="kirk-tripCard-bottom">
-            <div className="kirk-tripCard-driver">
-              <div className="kirk-tripCard-avatar">
-                <Avatar image={driver.avatarUrl} alt="" />
+            {driver && (
+              <div className="kirk-tripCard-driver">
+                <div className="kirk-tripCard-avatar">
+                  <Avatar image={driver.avatarUrl} />
+                </div>
+                <Text display={TextDisplayType.TITLE}>{driver.firstName}</Text>
               </div>
-              <Text display={TextDisplayType.TITLE}>{driver.firstName}</Text>
-            </div>
+            )}
+            {passengers && (
+              <ul className="kirk-tripCard-passengers">
+                {passengers.reverse().map(passenger => (
+                  <li className="kirk-tripCard-avatar" key={passenger.firstName}>
+                    <Avatar image={passenger.avatarUrl} isSmall />
+                  </li>
+                ))}
+              </ul>
+            )}
             {highlighted && (
               <Text className="kirk-tripCard-topText" display={TextDisplayType.TITLESTRONG}>
                 {highlighted}