+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+- **[NEW]** Added `passengers` option to `TripCard`
+- **[NEW]** Added `badge` option to `TripCard`
+- **[NEW]** Added `statusInformation` option to `TripCard`
+- **[NEW]** Added css shadow to `TripCard`
 - **[NEW]** Added `status` prop to `ItemAction`
 - **[FIX]** Fixed button with `CHECKED` status in `ButtonGroup`
 - **[DEV]** Added typescript check in pre-commit hook